commit: Added suggestion from the currentState to add all the hints to the resolved store. This is added because the hints were missing for few of the component values
